+++ benchmarks/rust/indicators/bench_ema.rs
@@ -2,6 +2,7 @@ use criterion::BenchmarkId;
 use rand::rngs::StdRng;
 use rand::{Rng, SeedableRng};
 use techalysis::indicators::ema::ema;
+use techalysis::types::Float;
 
 fn bench_ema(c: &mut criterion::Criterion) {
     let mut bench_group = c.benchmark_group("ema");
@@ -10,7 +11,7 @@ fn bench_ema(c: &mut criterion::Criterion) {
 
     for (len, period) in cases {
         let mut rng = StdRng::seed_from_u64(period as u64);
-        let data: Vec<f64> = (0..len).map(|_| rng.random_range(0.0..100.0)).collect();
+        let data: Vec<Float> = (0..len).map(|_| rng.random_range(0.0..100.0)).collect();
 
         bench_group.bench_with_input(
             BenchmarkId::new(format!("len={len}"), period),
